requirements.txt
fastapi
uvicorn
openai
sentence-transformers
faiss-cpu
numpy
pydantic
python-dotenv
google-generativeai